refactor: change context on repository
